+++ lua/primebrook/harpoon.lua
@@ -1,0 +1,9 @@
+require("harpoon").setup({
+	global_settings = {
+		save_on_toggle = true,
+		save_on_change = true,
+	},
+	menu = {
+		width = vim.api.nvim_win_get_width(0) - 4,
+	},
+})
