+++ lua/primebrook/treesitter.lua
@@ -1,0 +1,8 @@
+require("nvim-treesitter.configs").setup({
+	ensure_installed = { "c", "lua", "vim", "elixir", "python", "rust", "bash", "json" },
+	sync_install = false,
+	auto_install = true,
+	highlight = {
+		enable = true,
+	},
+})
