--- plugin/harpoon.lua
@@ -1,6 +1,0 @@
-local mark = require("harpoon.mark")
-local ui = require("harpoon.ui")
-
-vim.keymap.set("n", "<leader>a", mark.add_file, { noremap = true, silent = true })
-vim.keymap.set("n", "<C-e>", ui.toggle_quick_menu, { noremap = true, silent = true })
-vim.keymap.set("n", "<leader>ha", ui.toggle_quick_menu, { noremap = true, silent = true })
